requirements.txt
# Research Paper Writing Agents Requirements
# Core dependencies

# PDF parsing (required for Style Analyzer to learn from reference papers)
pdfplumber>=0.10.0  # Preferred PDF text extraction library
# Alternative: PyPDF2>=3.0.0  # Can be used as fallback

# Google Gen AI SDK for semantic role analysis (required for deep role analysis)
google-genai>=1.0.0  # Google Gen AI Python SDK - latest API for Gemini

# OpenAI API (optional, for OpenAI provider support)
openai>=1.0.0  # For GPT-based agents

# Optional: For enhanced functionality, you could add:
# langchain>=0.0.200  # For advanced LLM integration
# anthropic>=0.3.0  # For Claude-based agents

# Note: The system will work without PDF libraries if no reference papers are provided,
# but PDF libraries are required for Style Analyzer to learn from reference PDF papers.

